requirements.txt
fastapi
uvicorn
pillow
numpy
matplotlib
pytest
requests
python-dotenv
